#564 :truck: move wls exception type to api folder
