+++ wls-gui-wahllokalsystem/src/api/WLSException.ts
